fixed to long

+++ src/MyJetWallet.Fireblocks.Client/ApiClients.cs
@@ -9190,11 +9190,11 @@ namespace MyJetWallet.Fireblocks.Client
     
         /// <summary>Unix timestamp</summary>
         [Newtonsoft.Json.JsonProperty("createdAt", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public decimal CreatedAt { get; set; }
+        public long CreatedAt { get; set; }
     
         /// <summary>Unix timestamp</summary>
         [Newtonsoft.Json.JsonProperty("lastUpdated", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public decimal LastUpdated { get; set; }
+        public long LastUpdated { get; set; }
     
         [Newtonsoft.Json.JsonProperty("status", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
         [Newtonsoft.Json.JsonConverter(typeof(Newtonsoft.Json.Converters.StringEnumConverter))]
